Sixteenth Commit (1.9)
I hope this version does not piss some people off because, I removed
screens I was not using. If you are annoyed by this, I apologize in
advance. I am not a fan of course mode. Which is now gone. I also
removed the editor and jukebox. They are not needed. I cleaned the theme
up as much as I could and I even changed the sync song/step. (I kept it
similar to _fallback so hopefully, it won't throw files off sync... If
you find any problems, please let me know.